refactor(providers): keep environment merging in registry
